// src/coaching/enums/coachErrors.enum.ts
export enum CoachErrors {
  NO_COACH_ASSIGNED = 'NO_COACH_ASSIGNED',
  NO_COACH_PROFILE = 'NO_COACH_PROFILE',
  NOT_EXISTING_COACH = 'NOT_EXISTING_COACH',
  NOT_ENOUGH_COACHES = 'NOT_ENOUGH_COACHES',
  MAX_COACH_SUGGESTIONS_REACHED = 'MAX_COACH_SUGGESTIONS_REACHED',
  COACH_NOT_SUGGESTED = 'COACH_NOT_SUGGESTED',
  //
  COACH_NOTE_NOT_ALLOWED_TO_UPDATE = 'COACH_NOTE_NOT_ALLOWED_TO_UPDATE',
  COACH_NOTE_NOT_ALLOWED_TO_DELETE = 'COACH_NOTE_NOT_ALLOWED_TO_DELETE',
  COACH_NOTE_NOT_FOUND = 'COACH_NOTE_NOT_FOUND',
  //
  COACH_NOT_IN_COACHING_SESSION = 'COACH_NOT_IN_COACHING_SESSION',
}